Move GlobalEvent out of the RPC and add the FeedTransactionHandlerStrategy

+++ Rpc/Transactions/TransactionsWithAddressRequest.cs
@@ -8,7 +8,7 @@ public class TransactionsWithAddressRequest : CommandRequestBase
 
     public string Address { get; set; } = string.Empty;
 
-    public int LastHeightSynched { get; set; }
+    public double LastHeightSynched { get; set; }
 
     public TransactionsWithAddressRequest()
     {
